Restructured the Blurbs into BlurbContainer and BlurbText to fix left-to-right alignment  + horizontal/vertical centering issues. Fixed some @media bugs that caused the Blurbs to have a strange elongated shape on high screen resolutions. Changed font sizes/weights/styles of headers so that they are uniform across pages.
Created OurStory.js + OurStory.css for better organization. Rearranged css selectors in About.css, OurStory.css, and What.css so that the selectors that apply to all of these pages are in About.css, and that the selectors that apply only to one page are in the css stylesheets for their respective pages.

Todo: Although ALIGNMENT is now much better on most screen resolutions, now the SIZES (font, blurb height, image, etc.) need to be adjusted to display better on large screens.